requirements.txt
asgiref==3.9.1
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
cryptography==45.0.5
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==21.2.0
idna==3.11
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
